requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
footprint_mongoengine==0.5.4
idna==3.4
mongoengine==0.27.0
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
